requirements.txt
beautifulsoup4
fastapi
uvicorn
requests
python-dotenv
langchain
chromadb
google-generativeai
pydantic
aiohttp
lxml
html5lib
jinja2
httpx
validators
tiktoken
qdrant-client
typing-extensions
asyncio